feat(metadata): add metadata model and utilities
This change introduces the `Metadata` struct and related utility functions to
handle the metadata of a resource in the Sonr ecosystem. The metadata includes
information such as the public key, path, accumulators, address, chain ID,
validator address, RPC endpoint, API endpoint, IPFS endpoint, peer ID, and
supported denominations.

The `Accumulators` type is a map of accumulators, which can be used to store
and retrieve the accumulated data of secret data associated with a resource.
The `Marshal` and `Unmarshal` methods are provided to handle the JSON
encoding/decoding of the `Accumulators` and `Metadata` types.

Additionally, the `Save` method is added to the `Metadata` struct to write the
JSON-encoded metadata to a provided file.